feat: inline black box function logic from `common`
